Revert "First Stab at migrating many of the Double objects to Float's"
This reverts commit 3f6db598a75958fc45aa9667ec6567b638c775a6.

Conflicts:
	Common/Controls/ColorPicker/ColorPicker.cs
	Common/Controls/ColorPicker/ColorSelectionFader.cs
	Common/Controls/ColorPicker/ColorSelectionModules_HSV.cs
	Common/Controls/ColorPicker/ColorSelectionModules_LAB.cs
	Common/Controls/ColorPicker/ColorSelectionModules_RGB.cs
	Common/Controls/ColorPicker/ColorSelectionPlane.cs
	Modules/App/ColorGradients/ColorGradient.cs
	Modules/App/ColorGradients/GradientEdit.cs
	Modules/App/ColorGradients/GradientEditPanel.cs
	Modules/App/Curves/Curve.cs
	Modules/Effect/Chase/Chase.cs
	Modules/Effect/Spin/Spin.cs
	Modules/Effect/Twinkle/Twinkle.cs
	Modules/PostFilter/DimmingCurve/DimmingCurveModule.cs
	Vixen.System/Data/Value/LightingValue.cs
	Vixen.System/Data/Value/PositionValue.cs
	Vixen.System/Interpolator/Interpolator.cs
	Vixen.System/Interpolator/LightingValueInterpolator.cs
	Vixen.System/Interpolator/PositionValueInterpolator.cs
	Vixen.System/Interpolator/RGBValueInterpolator.cs
	Vixen.System/Interpolator/StaticValueInterpolator.cs

+++ Common/Controls/ColorPicker/ColorPicker.cs
@@ -658,8 +658,8 @@ namespace Common.Controls.ColorManagement.ColorPicker
 				e.Handled = true;
 				return;
 			}
-			float value;
-			if (!float.TryParse(((TextBox) sender).Text,
+			double value;
+			if (!double.TryParse(((TextBox) sender).Text,
 			                     System.Globalization.NumberStyles.Integer,
 			                     null, out value)) return;
 
@@ -667,17 +667,17 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 			if (sender == tbHSV_H) {
 				HSV chsv = HSV.FromRGB(InternalColor.ToRGB());
-				chsv.H = (float)value/360.0f;
+				chsv.H = value/360.0;
 				InternalColor = XYZ.FromRGB(chsv.ToRGB());
 			}
 			else if (sender == tbHSV_S) {
 				HSV chsv = HSV.FromRGB(InternalColor.ToRGB());
-				chsv.S = (float)value/100.0f;
+				chsv.S = value/100.0;
 				InternalColor = XYZ.FromRGB(chsv.ToRGB());
 			}
 			else if (sender == tbHSV_V) {
 				HSV chsv = HSV.FromRGB(InternalColor.ToRGB());
-				chsv.V = (float)value/100.0f;
+				chsv.V = value/100.0;
 				InternalColor = XYZ.FromRGB(chsv.ToRGB());
 			}
 				#endregion
@@ -686,14 +686,14 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			else if (_mode == Mode.HSV_RGB) {
 				RGB crgb = InternalColor.ToRGB();
 				if (sender == tbSecond_1) {
-					crgb.R = value/255.0f;
+					crgb.R = value/255.0;
 				}
 				else if (sender == tbSecond_2) {
-					crgb.G = value/255.0f;
+					crgb.G = value/255.0;
 				}
 				else //sender==tbSecond_3
 				{
-					crgb.B = value/255.0f;
+					crgb.B = value/255.0;
 				}
 				InternalColor = XYZ.FromRGB(crgb);
 			}
@@ -816,7 +816,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			{
 				if (LockValue_V) {
 					HSV temp = HSV.FromRGB(value.ToRGB());
-					temp.V = 1.0f;
+					temp.V = 1.0;
 					value = XYZ.FromRGB(temp.ToRGB());
 				}
 				_color = value;
