+++ Common/Controls/ColorPicker/ColorSelectionModules_LAB.cs
@@ -46,8 +46,8 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			                             ImageLockMode.WriteOnly, PixelFormat.Format32bppArgb);
 			ColorBgra* scan0 = (ColorBgra*) bd.Scan0;
 			//draw fader image
-			LAB curr = new LAB(0.0f, _color.a, _color.b);
-			float delta_L = 100.0f/(float) bd.Width;
+			LAB curr = new LAB(0.0, _color.a, _color.b);
+			double delta_L = 100.0/(double) bd.Width;
 			for (int x = 0; x < bd.Width; x++, scan0++,curr.L += delta_L) {
 				scan0[0] = ColorBgra.FromArgb(curr.ToXYZ().ToRGB());
 			}
@@ -57,12 +57,12 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdateFaderPosition(ColorSelectionFader fader)
 		{
-			fader.Position = _color.L/100.0f;
+			fader.Position = _color.L/100.0;
 		}
 
 		protected override void OnFaderScroll(ColorSelectionFader fader)
 		{
-			float newL = fader.Position*100.0f;
+			double newL = fader.Position*100.0;
 			if (newL == _color.L) return;
 			_color.L = newL;
 			RaiseColorChanged();
@@ -81,11 +81,11 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			                             ImageLockMode.WriteOnly, PixelFormat.Format32bppArgb);
 			ColorBgra* scan0 = (ColorBgra*) bd.Scan0;
 			//draw plane image
-			LAB curr = new LAB(_color.L, -128.0f, 127.0f);
-			float
-				delta_a = 255.0f/(float) bd.Width,
-				delta_b = -255.0f/(float) bd.Height;
-			for (int y = 0; y < bd.Height; y++, curr.a = -128.0f, curr.b += delta_b) {
+			LAB curr = new LAB(_color.L, -128.0, 127.0);
+			double
+				delta_a = 255.0/(double) bd.Width,
+				delta_b = -255.0/(double) bd.Height;
+			for (int y = 0; y < bd.Height; y++, curr.a = -128.0, curr.b += delta_b) {
 				for (int x = 0; x < bd.Width; x++, scan0++,curr.a += delta_a) {
 					scan0[0] = ColorBgra.FromArgb(curr.ToXYZ().ToRGB());
 				}
@@ -96,14 +96,14 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdatePlanePosition(ColorSelectionPlane plane)
 		{
-			plane.SetPosition((_color.a + 128.0f)/255.0f,
-			                  (127.0f - _color.b)/255.0f);
+			plane.SetPosition((_color.a + 128.0)/255.0,
+			                  (127.0 - _color.b)/255.0);
 		}
 
 		protected override void OnPlaneScroll(ColorSelectionPlane plane)
 		{
-			float newa = plane.PositionX*255.0f - 128.0f,
-			       newb = 127.0f - plane.PositionY*255.0f;
+			double newa = plane.PositionX*255.0 - 128.0,
+			       newb = 127.0 - plane.PositionY*255.0;
 			if (newa == _color.a &&
 			    newb == _color.b) return;
 			_color.a = newa;
@@ -126,8 +126,8 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			                             ImageLockMode.WriteOnly, PixelFormat.Format32bppArgb);
 			ColorBgra* scan0 = (ColorBgra*) bd.Scan0;
 			//draw fader image
-			LAB curr = new LAB(_color.L, -128.0f, _color.b);
-			float delta_a = 255.0f/(float) bd.Width;
+			LAB curr = new LAB(_color.L, -128.0, _color.b);
+			double delta_a = 255.0/(double) bd.Width;
 			for (int x = 0; x < bd.Width; x++, scan0++,curr.a += delta_a) {
 				scan0[0] = ColorBgra.FromArgb(curr.ToXYZ().ToRGB());
 			}
@@ -137,12 +137,12 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdateFaderPosition(ColorSelectionFader fader)
 		{
-			fader.Position = (_color.a + 128.0f)/255.0f;
+			fader.Position = (_color.a + 128.0)/255.0;
 		}
 
 		protected override void OnFaderScroll(ColorSelectionFader fader)
 		{
-			float newa = fader.Position*255.0f - 128.0f;
+			double newa = fader.Position*255.0 - 128.0;
 			if (newa == _color.b) return;
 			_color.a = newa;
 			RaiseColorChanged();
@@ -161,11 +161,11 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			                             ImageLockMode.WriteOnly, PixelFormat.Format32bppArgb);
 			ColorBgra* scan0 = (ColorBgra*) bd.Scan0;
 			//draw plane image
-			LAB curr = new LAB(100.0f, _color.a, -128.0f);
-			float
-				delta_L = -100.0f/(float) bd.Height,
-				delta_b = 255.0f/(float) bd.Width;
-			for (int y = 0; y < bd.Height; y++, curr.b = -128.0f, curr.L += delta_L) {
+			LAB curr = new LAB(100.0, _color.a, -128.0);
+			double
+				delta_L = -100.0/(double) bd.Height,
+				delta_b = 255.0/(double) bd.Width;
+			for (int y = 0; y < bd.Height; y++, curr.b = -128.0, curr.L += delta_L) {
 				for (int x = 0; x < bd.Width; x++, scan0++,curr.b += delta_b) {
 					scan0[0] = ColorBgra.FromArgb(curr.ToXYZ().ToRGB());
 				}
@@ -176,14 +176,14 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdatePlanePosition(ColorSelectionPlane plane)
 		{
-			plane.SetPosition((_color.b + 128.0f)/255.0f,
-			                  1.0f - _color.L/100.0f);
+			plane.SetPosition((_color.b + 128.0)/255.0,
+			                  1.0 - _color.L/100.0);
 		}
 
 		protected override void OnPlaneScroll(ColorSelectionPlane plane)
 		{
-			float newb = plane.PositionX*255.0f - 128.0f,
-			       newL = (1.0f - plane.PositionY)*100.0f;
+			double newb = plane.PositionX*255.0 - 128.0,
+			       newL = (1.0 - plane.PositionY)*100.0;
 			if (newb == _color.b &&
 			    newL == _color.L) return;
 			_color.b = newb;
@@ -206,8 +206,8 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			                             ImageLockMode.WriteOnly, PixelFormat.Format32bppArgb);
 			ColorBgra* scan0 = (ColorBgra*) bd.Scan0;
 			//draw fader image
-			LAB curr = new LAB(_color.L, _color.a, -128.0f);
-			float delta_b = 255.0f/(float) bd.Width;
+			LAB curr = new LAB(_color.L, _color.a, -128.0);
+			double delta_b = 255.0/(double) bd.Width;
 			for (int x = 0; x < bd.Width; x++, scan0++,curr.b += delta_b) {
 				scan0[0] = ColorBgra.FromArgb(curr.ToXYZ().ToRGB());
 			}
@@ -217,12 +217,12 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdateFaderPosition(ColorSelectionFader fader)
 		{
-			fader.Position = (_color.b + 128.0f)/255.0f;
+			fader.Position = (_color.b + 128.0)/255.0;
 		}
 
 		protected override void OnFaderScroll(ColorSelectionFader fader)
 		{
-			float newb = fader.Position*255.0f - 128.0f;
+			double newb = fader.Position*255.0 - 128.0;
 			if (newb == _color.b) return;
 			_color.b = newb;
 			RaiseColorChanged();
@@ -241,11 +241,11 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			                             ImageLockMode.WriteOnly, PixelFormat.Format32bppArgb);
 			ColorBgra* scan0 = (ColorBgra*) bd.Scan0;
 			//draw plane image
-			LAB curr = new LAB(100.0f, -128.0f, _color.b);
-			float
-				delta_L = -100.0f/(float) bd.Height,
-				delta_a = 255.0f/(float) bd.Width;
-			for (int y = 0; y < bd.Height; y++, curr.a = -128.0f, curr.L += delta_L) {
+			LAB curr = new LAB(100.0, -128.0, _color.b);
+			double
+				delta_L = -100.0/(double) bd.Height,
+				delta_a = 255.0/(double) bd.Width;
+			for (int y = 0; y < bd.Height; y++, curr.a = -128.0, curr.L += delta_L) {
 				for (int x = 0; x < bd.Width; x++, scan0++,curr.a += delta_a) {
 					scan0[0] = ColorBgra.FromArgb(curr.ToXYZ().ToRGB());
 				}
@@ -256,14 +256,14 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdatePlanePosition(ColorSelectionPlane plane)
 		{
-			plane.SetPosition((_color.a + 128.0f)/255.0f,
-			                  1.0f - _color.L/100.0f);
+			plane.SetPosition((_color.a + 128.0)/255.0,
+			                  1.0 - _color.L/100.0);
 		}
 
 		protected override void OnPlaneScroll(ColorSelectionPlane plane)
 		{
-			float newa = plane.PositionX*255.0f - 128.0f,
-			       newL = (1.0f - plane.PositionY)*100.0f;
+			double newa = plane.PositionX*255.0 - 128.0,
+			       newL = (1.0 - plane.PositionY)*100.0;
 			if (newa == _color.a &&
 			    newL == _color.L) return;
 			_color.a = newa;
